+++ packages/orm/angel_migration_runner/lib/src/mysql/runner.dart
@@ -54,8 +54,11 @@ class MySqlMigrationRunner implements MigrationRunner {
   @override
   Future up() async {
     await _init();
-    var r = await connection.execute('SELECT path from migrations;');
-    var existing = r.rows.cast<String>(); //.expand((x) => x).cast<String>();
+    var result = await connection.execute('SELECT path from migrations;');
+    var existing = <String>[];
+    if (result.rows.isNotEmpty) {
+      existing = result.rows.first.assoc().values.cast<String>().toList();
+    }
     var toRun = <String>[];
 
     migrations.forEach((k, v) {
@@ -63,11 +66,14 @@ class MySqlMigrationRunner implements MigrationRunner {
     });
 
     if (toRun.isNotEmpty) {
-      var r = await connection.execute('SELECT MAX(batch) from migrations;');
-      var rTmp = r.rows.first; //r.toList();
-      var curBatch =
-          int.tryParse(rTmp.colAt(0) ?? "0") ?? 0; //(rTmp[0][0] ?? 0) as int;
-      var batch = curBatch + 1;
+      var result =
+          await connection.execute('SELECT MAX(batch) from migrations;');
+      var curBatch = 0;
+      if (result.rows.isNotEmpty) {
+        var firstRow = result.rows.first;
+        curBatch = int.tryParse(firstRow.colAt(0) ?? "0") ?? 0;
+      }
+      curBatch++;
 
       for (var k in toRun) {
         var migration = migrations[k]!;
@@ -77,16 +83,13 @@ class MySqlMigrationRunner implements MigrationRunner {
         await schema.run(connection).then((_) async {
           var result = await connection
               .execute(
-                  "INSERT INTO MIGRATIONS (batch, path) VALUES ($batch, '$k')")
+                  "INSERT INTO migrations (batch, path) VALUES ($curBatch, '$k')")
               .catchError((e) {
             _log.severe('Failed to insert into "migrations" table.', e);
           });
 
           return result.affectedRows.toInt();
         });
-        //return connection.execute(
-        //    'INSERT INTO MIGRATIONS (batch, path) VALUES ($batch, \'$k\');');
-
       }
     } else {
       _log.warning('Nothing to add into "migrations" table.');
@@ -97,14 +100,18 @@ class MySqlMigrationRunner implements MigrationRunner {
   Future rollback() async {
     await _init();
 
-    var r = await connection.execute('SELECT MAX(batch) from migrations;');
-    var rTmp = r.rows.first; //r.toList();
-    var curBatch =
-        int.tryParse(rTmp.colAt(0) ?? "0") ?? 0; //(rTmp[0][0] ?? 0) as int;
-
-    r = await connection
+    var result = await connection.execute('SELECT MAX(batch) from migrations;');
+    var curBatch = 0;
+    if (result.rows.isNotEmpty) {
+      var firstRow = result.rows.first;
+      curBatch = int.tryParse(firstRow.colAt(0) ?? "0") ?? 0;
+    }
+    result = await connection
         .execute('SELECT path from migrations WHERE batch = $curBatch;');
-    var existing = r.rows.cast<String>(); //r.expand((x) => x).cast<String>();
+    var existing = <String>[];
+    if (result.rows.isNotEmpty) {
+      existing = result.rows.first.assoc().values.cast<String>().toList();
+    }
     var toRun = <String>[];
 
     migrations.forEach((k, v) {
@@ -130,9 +137,13 @@ class MySqlMigrationRunner implements MigrationRunner {
   @override
   Future reset() async {
     await _init();
-    var r = await connection
+    var result = await connection
         .execute('SELECT path from migrations ORDER BY batch DESC;');
-    var existing = r.rows.cast<String>(); //r.expand((x) => x).cast<String>();
+    var existing = <String>[];
+    if (result.rows.isNotEmpty) {
+      var firstRow = result.rows.first;
+      existing = firstRow.assoc().values.cast<String>().toList();
+    }
     var toRun = existing.where(migrations.containsKey).toList();
 
     if (toRun.isNotEmpty) {